[refactor] : State 객체를 구현한 추상클래스(SingleBaseState, DoubleBaseState)를 두어 중복된 코드 제거, 오버라이딩을 통한 State별 로직처리
